// apps/frontend/src/payload-types.ts
export interface Topic {
  id: string;
  title: string;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface InternalDocument {
  id: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Product {
  id: string;
  title: string;
  amount: number;
  is_enabled: boolean;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface User {
  id: string;
  title?: ('' | 'Sr.' | 'Sr.ª' | 'Dr.' | 'Dr.ª' | 'Prof.' | 'Prof.ª' | 'Exmo.' | 'Exmo.ª') | null;
  first_name?: string | null;
  last_name?: string | null;
  tax_id: string;
  medical_id?: number | null;
  birthday?: string | null;
  member_since?: string | null;
  send_newsletter?: boolean | null;
  phone?: string | null;
  address_1?: string | null;
  address_2?: string | null;
  postal_code?: string | null;
  city?: string | null;
  country?: string | null;
  workplace_primary?: string | null;
  workplace_secondary?: string | null;
  subscribed_sections?:
    | (
        | 'colposcopia_patologia_tracto_genital_inferior'
        | 'endoscopia_ginecologica'
        | 'ginecologia_oncologica'
        | 'menopausa'
        | 'uroginecologia'
      )[]
    | null;
  billing_name?: string | null;
  billing_tax_id?: string | null;
  billing_address_1?: string | null;
  billing_address_2?: string | null;
  billing_postal_code?: string | null;
  billing_city?: string | null;
  stripe_id?: string | null;
  transactions?:
    | {
        doc_type?: ('invoice' | 'credit_note') | null;
        doc_date?: string | null;
        doc_amount?: number | null;
        doc_number?: string | null;
        doc_id?: string | null;
        doc_system_time?: string | null;
        associated_products?: (string | Product)[] | null;
        id?: string | null;
      }[]
    | null;
  account_status?: ('active' | 'pending' | 'dormant') | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "videos".
 */
export interface Video {
  id: string;
  title?: string | null;
  status: 'draft' | 'in_review' | 'approved' | 'rejected';
  video_file?: (string | null) | VideoFile;
  declaration_file?: (string | null) | InternalDocument;
  authors?: string | null;
  section?:
    | (
        | 'geral'
        | 'colposcopia_patologia_tracto_genital_inferior'
        | 'endoscopia_ginecologica'
        | 'ginecologia_oncologica'
        | 'menopausa'
        | 'uroginecologia'
      )
    | null;
  introduction?: string | null;
  description?: string | null;
  publisher?: (string | null) | User;
  views?: number | null;
  is_featured?: boolean | null;
  topics?: (string | Topic)[] | null;
  publishedAt: string;
  featured_image?: (string | null) | Media;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "video-files".
 */
export interface VideoFile {
  id: string;
  duration?: number | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}